requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
passlib==1.7.1
pymssql==2.1.4
mysqlclient==1.4.2.post1
SQLAlchemy==1.3.1

